Modified Direction, Position, and Moving Game Element to use floats again instead of integers.
Modified Game to capture canvas dimensions on initialization and draw screen elements based on percentages of the dimensions again.  Reduced the maximum speed to accommodate calculation of position in game loop.

Modified Game Thread to pass the canvas to Game on initialization instead of its width.  Game gets the dimensions from it internally.

+++ app/src/main/java/com/twilightcitizen/whackapede/gameElements/Position.java
@@ -9,12 +9,12 @@ package com.twilightcitizen.whackapede.gameElements;
 
 public class Position {
     // Components along the X and Y axes.
-    private final int x;
-    private final int y;
+    private final float x;
+    private final float y;
 
-    public Position( int x, int y ) { this.x = x; this.y = y; }
+    public Position( float x, float y ) { this.x = x; this.y = y; }
 
     // Components along the X and Y axes are read-only.
-    public int getX() { return x; }
-    public int getY() { return y; }
+    public float getX() { return x; }
+    public float getY() { return y; }
 }
